Split twitchEventSubSubscriptions.ts to fix qlty's high-total-complexity flag
The qlty bot flagged src/twitch/eventsub/twitchEventSubSubscriptions.ts
(qlty:file-complexity, count=50) — the actual target of the earlier
complexity concern, not twitchEventSubHandler.ts as previously assumed.

Extracted two self-contained pieces of branching logic into their own
modules, matching the codebase's existing pattern of one concern per file:

- twitchEventSubSubscriptionGroups.ts: the SUBSCRIPTION_GROUPS table (5
  groups, each with a configGate closure and specs builder),
  isGroupEnabled, and getAlertTypesCoveredBySubscriptionGroups.
- twitchEventSubDispatch.ts: the streamer map, notificationHandlers
  routing table, dispatchNotification, and handleRevocation.

twitchEventSubSubscriptions.ts re-exports the moved public functions
(dispatchNotification, handleRevocation, removeStreamerFromMap,
getAlertTypesCoveredBySubscriptionGroups) so every other consumer
(twitchEventSubConnection.ts, eventsubCallback.ts, tests) needed no
changes. Corresponding test blocks moved into co-located test files for
the two new modules; no behavior change.
