+++ src/twitch/eventsub/twitchEventSubSubscriptions.test.ts
@@ -4,32 +4,16 @@ const { logMock } = vi.hoisted(() => ({
   logMock: { info: vi.fn(), warn: vi.fn(), error: vi.fn(), debug: vi.fn() },
 }));
 vi.mock('../../shared/logger', () => ({ createLogger: () => logMock }));
-// Pulled in transitively via '../../db/alertConfig' (for the real ALERT_EVENT_TYPES below)
-// importing './pool', which reads real env vars at module load time — an empty mock keeps
-// that side effect out.
 vi.mock('../../shared/config', () => ({}));
-vi.mock('../../db', async () => {
-  // Reuses the real ALERT_EVENT_TYPES (rather than a hard-coded copy) so this file's
-  // exhaustiveness test can't silently go stale if a new alert type is ever added there.
-  const { ALERT_EVENT_TYPES } = await vi.importActual<typeof import('../../db/alertConfig')>('../../db/alertConfig');
-  return {
-    getAllEventSubStreamers: vi.fn(),
-    clearStreamerToken: vi.fn().mockResolvedValue(undefined),
-    getEnabledAlertEventTypesBatch: vi.fn().mockResolvedValue(new Map()),
-    ALERT_EVENT_TYPES,
-    DEFAULT_EVENT_CONFIG: {
-      follow_enabled: false,
-      follow_message: 'Thanks {display_name} for the follow!',
-      sub_enabled: false,
-      sub_message: 'Thanks {display_name} for subscribing! ({tier_name})',
-      resub_message: 'Thanks {display_name} for {months} months! ({tier_name})',
-      giftsub_message: '{gifter_display} gifted {count} sub(s) to the community!',
-      raid_enabled: false,
-      raid_message: 'Welcome raiders from {from_display}! Thank you for the {viewers} person raid!',
-      raid_shoutout_enabled: false,
-    },
-  };
-});
+// './twitchEventSubSubscriptions' re-exports dispatchNotification/handleRevocation/
+// removeStreamerFromMap from './twitchEventSubDispatch', which in turn imports these two real
+// modules — mocked here purely so that transitive import chain resolves, even though this
+// file's own tests exercise subscribeForStreamer/loadStreamersForEventSub, not dispatch.
+vi.mock('../../db', () => ({
+  getAllEventSubStreamers: vi.fn(),
+  clearStreamerToken: vi.fn().mockResolvedValue(undefined),
+  getEnabledAlertEventTypesBatch: vi.fn().mockResolvedValue(new Map()),
+}));
 vi.mock('../twitchApi', () => ({ getUsers: vi.fn() }));
 vi.mock('../twitchChannelMembership', () => ({ getActiveChannels: vi.fn().mockReturnValue(new Set<string>()) }));
 vi.mock('../twitchChannelName', () => ({ normalizeTwitchChannelName: vi.fn((n: string) => n.toLowerCase()) }));
@@ -57,21 +41,11 @@ import {
   clearAuthFailedSubs,
   loadStreamersForEventSub,
   subscribeForStreamer,
-  dispatchNotification,
-  removeStreamerFromMap,
-  handleRevocation,
-  getAlertTypesCoveredBySubscriptionGroups,
 } from './twitchEventSubSubscriptions';
-import {
-  getAllEventSubStreamers, clearStreamerToken, getEnabledAlertEventTypesBatch, ALERT_EVENT_TYPES, DEFAULT_EVENT_CONFIG,
-} from '../../db';
+import { getAllEventSubStreamers, getEnabledAlertEventTypesBatch } from '../../db';
 import { getValidToken, createEventSubSubscription, listEventSubSubscriptions, deleteEventSubSubscription, TwitchAuthError } from './twitchApiEventSub';
 import { getUsers } from '../twitchApi';
 import { getActiveChannels } from '../twitchChannelMembership';
-import {
-  handleStreamOnline, handleStreamOffline, handleChannelUpdate,
-  handleFollow, handleSub, handleResub, handleGiftSub, handleRaid, handleRedemption,
-} from './twitchEventSubHandler';
 
 // ---------------------------------------------------------------------------
 // hasAuthFailedSubs / clearAuthFailedSubs
@@ -526,42 +500,6 @@ describe('subscribeForStreamer', () => {
 });
 
 // ---------------------------------------------------------------------------
-// dispatchNotification — stream.online / stream.offline routing
-// ---------------------------------------------------------------------------
-describe('dispatchNotification routes stream.online/offline', () => {
-  beforeEach(async () => {
-    vi.clearAllMocks();
-    clearAuthFailedSubs('liveStreamer');
-    vi.mocked(getActiveChannels).mockReturnValue(new Set(['livestreamer']));
-    vi.mocked(createEventSubSubscription).mockResolvedValue('sub-id');
-    vi.mocked(listEventSubSubscriptions).mockResolvedValue([]);
-    // Populate streamerMap with a known config so dispatchNotification doesn't early-exit.
-    await subscribeForStreamer('sess-dispatch', {
-      uid: 'uid-dispatch',
-      token: 'tok-dispatch',
-      name: 'liveStreamer',
-      config: { follow_enabled: false, sub_enabled: false, raid_enabled: false } as any,
-      streamerId: 30,
-    });
-  });
-
-  it('calls handleStreamOnline for a stream.online notification', () => {
-    dispatchNotification('stream.online', {}, { broadcaster_user_id: 'uid-dispatch' });
-    expect(handleStreamOnline).toHaveBeenCalledWith('liveStreamer');
-  });
-
-  it('calls handleStreamOffline for a stream.offline notification', () => {
-    dispatchNotification('stream.offline', {}, { broadcaster_user_id: 'uid-dispatch' });
-    expect(handleStreamOffline).toHaveBeenCalledWith('liveStreamer');
-  });
-
-  it('calls handleChannelUpdate for a channel.update notification', () => {
-    dispatchNotification('channel.update', {}, { broadcaster_user_id: 'uid-dispatch' });
-    expect(handleChannelUpdate).toHaveBeenCalledWith('liveStreamer');
-  });
-});
-
-// ---------------------------------------------------------------------------
 // subscribeForStreamer — sub_enabled subscription creation
 // ---------------------------------------------------------------------------
 describe('subscribeForStreamer with sub_enabled', () => {
@@ -606,128 +544,6 @@ describe('subscribeForStreamer with sub_enabled', () => {
     expect(createEventSubSubscription).not.toHaveBeenCalledWith(
       'channel.subscribe', expect.anything(), expect.anything(), expect.anything(), expect.anything(),
     );
-  });
-});
-
-// ---------------------------------------------------------------------------
-// dispatchNotification — remaining notification types
-// ---------------------------------------------------------------------------
-describe('dispatchNotification routes chat-alert and redemption events', () => {
-  beforeEach(async () => {
-    vi.clearAllMocks();
-    clearAuthFailedSubs('alertStreamer');
-    vi.mocked(getActiveChannels).mockReturnValue(new Set(['alertstreamer']));
-    vi.mocked(createEventSubSubscription).mockResolvedValue('sub-id');
-    vi.mocked(listEventSubSubscriptions).mockResolvedValue([]);
-    await subscribeForStreamer('sess-alert', {
-      uid: 'uid-alert',
-      token: 'tok-alert',
-      name: 'alertStreamer',
-      config: { follow_enabled: true, sub_enabled: true, raid_enabled: true } as any,
-      streamerId: 50,
-    });
-  });
-
-  it('routes channel.follow to handleFollow', () => {
-    dispatchNotification('channel.follow', { user_name: 'follower' }, { broadcaster_user_id: 'uid-alert' });
-    expect(handleFollow).toHaveBeenCalledWith('alertStreamer', { user_name: 'follower' }, expect.anything(), 50);
-  });
-
-  it('routes channel.subscribe to handleSub', () => {
-    dispatchNotification('channel.subscribe', { tier: '1000' }, { broadcaster_user_id: 'uid-alert' });
-    expect(handleSub).toHaveBeenCalledWith('alertStreamer', { tier: '1000' }, expect.anything(), 50);
-  });
-
-  it('routes channel.subscription.message to handleResub', () => {
-    dispatchNotification('channel.subscription.message', { cumulative_months: 5 }, { broadcaster_user_id: 'uid-alert' });
-    expect(handleResub).toHaveBeenCalledWith('alertStreamer', { cumulative_months: 5 }, expect.anything(), 50);
-  });
-
-  it('routes channel.subscription.gift to handleGiftSub', () => {
-    dispatchNotification('channel.subscription.gift', { total: 3 }, { broadcaster_user_id: 'uid-alert' });
-    expect(handleGiftSub).toHaveBeenCalledWith('alertStreamer', { total: 3 }, expect.anything(), 50);
-  });
-
-  it('routes channel.raid to handleRaid using to_broadcaster_user_id', () => {
-    dispatchNotification('channel.raid', { from_broadcaster_user_name: 'raider' }, { to_broadcaster_user_id: 'uid-alert' });
-    expect(handleRaid).toHaveBeenCalledWith('alertStreamer', { from_broadcaster_user_name: 'raider' }, expect.anything(), 50);
-  });
-
-  it('routes a channel-points redemption to handleRedemption with the streamerId', () => {
-    dispatchNotification(
-      'channel.channel_points_custom_reward_redemption.add',
-      { reward: { id: 'r1' } },
-      { broadcaster_user_id: 'uid-alert' },
-    );
-    expect(handleRedemption).toHaveBeenCalledWith('alertStreamer', { reward: { id: 'r1' } }, expect.anything(), 50);
-  });
-
-  it('logs a warning and does nothing for an unsupported notification type', () => {
-    dispatchNotification('channel.cheer', {}, { broadcaster_user_id: 'uid-alert' });
-    expect(logMock.warn).toHaveBeenCalledWith('Unsupported EventSub notification type: channel.cheer');
-    expect(handleFollow).not.toHaveBeenCalled();
-  });
-
-  it('logs an error when the routed handler rejects', async () => {
-    vi.mocked(handleFollow).mockRejectedValueOnce(new Error('handler exploded'));
-
-    dispatchNotification('channel.follow', {}, { broadcaster_user_id: 'uid-alert' });
-    await new Promise((resolve) => setImmediate(resolve));
-
-    expect(logMock.error).toHaveBeenCalledWith('channel.follow handler error:', expect.any(Error));
-  });
-});
-
-// ---------------------------------------------------------------------------
-// dispatchNotification — streamer with no streamer_event_config row
-// ---------------------------------------------------------------------------
-describe('dispatchNotification with a null streamer config (alert-only streamer)', () => {
-  beforeEach(async () => {
-    vi.clearAllMocks();
-    clearAuthFailedSubs('alertOnlyStreamer');
-    vi.mocked(getActiveChannels).mockReturnValue(new Set(['alertonlystreamer']));
-    vi.mocked(createEventSubSubscription).mockResolvedValue('sub-id');
-    vi.mocked(listEventSubSubscriptions).mockResolvedValue([]);
-    await subscribeForStreamer('sess-alert-only', {
-      uid: 'uid-alert-only',
-      token: 'tok-alert-only',
-      name: 'alertOnlyStreamer',
-      config: null,
-      streamerId: 60,
-    });
-  });
-
-  it('still routes the notification, falling back to DEFAULT_EVENT_CONFIG instead of dropping it', () => {
-    dispatchNotification('channel.follow', { user_name: 'follower' }, { broadcaster_user_id: 'uid-alert-only' });
-    expect(handleFollow).toHaveBeenCalledWith('alertOnlyStreamer', { user_name: 'follower' }, DEFAULT_EVENT_CONFIG, 60);
-  });
-});
-
-// ---------------------------------------------------------------------------
-// removeStreamerFromMap
-// ---------------------------------------------------------------------------
-describe('removeStreamerFromMap', () => {
-  beforeEach(async () => {
-    vi.clearAllMocks();
-    clearAuthFailedSubs('removable');
-    vi.mocked(getActiveChannels).mockReturnValue(new Set(['removable']));
-    vi.mocked(createEventSubSubscription).mockResolvedValue('sub-id');
-    vi.mocked(listEventSubSubscriptions).mockResolvedValue([]);
-    await subscribeForStreamer('sess-rm', {
-      uid: 'uid-rm',
-      token: 'tok-rm',
-      name: 'removable',
-      config: { follow_enabled: false, sub_enabled: false, raid_enabled: false } as any,
-      streamerId: 60,
-    });
-  });
-
-  it('removes the streamer so dispatchNotification no longer routes to it', () => {
-    removeStreamerFromMap('uid-rm');
-
-    dispatchNotification('stream.online', {}, { broadcaster_user_id: 'uid-rm' });
-
-    expect(handleStreamOnline).not.toHaveBeenCalled();
   });
 });
 
@@ -789,86 +605,5 @@ describe('error handling in subscription setup', () => {
       'Failed to subscribe to channel.follow for errStreamer:', expect.any(Error),
     );
     expect(hasAuthFailedSubs('errStreamer')).toBe(false);
-  });
-});
-
-// ---------------------------------------------------------------------------
-// handleRevocation
-// ---------------------------------------------------------------------------
-describe('handleRevocation', () => {
-  beforeEach(async () => {
-    vi.clearAllMocks();
-    clearAuthFailedSubs('revokedStreamer');
-    vi.mocked(getActiveChannels).mockReturnValue(new Set(['revokedstreamer']));
-    vi.mocked(createEventSubSubscription).mockResolvedValue('sub-id');
-    vi.mocked(listEventSubSubscriptions).mockResolvedValue([]);
-    vi.mocked(clearStreamerToken).mockResolvedValue(true);
-    await subscribeForStreamer('sess-revoke', {
-      uid: 'uid-revoke',
-      token: 'tok-revoke',
-      name: 'revokedStreamer',
-      config: { follow_enabled: false, sub_enabled: false, raid_enabled: false } as any,
-      streamerId: 100,
-    });
-  });
-
-  it('logs the revocation but does not clear a token for an unknown broadcaster', () => {
-    handleRevocation({ type: 'channel.follow', status: 'authorization_revoked', condition: { broadcaster_user_id: 'unknown-uid' } });
-
-    expect(logMock.warn).toHaveBeenCalledWith('Subscription revoked: type=channel.follow status=authorization_revoked');
-    expect(clearStreamerToken).not.toHaveBeenCalled();
-  });
-
-  it('clears the token when status is authorization_revoked for a known broadcaster', () => {
-    handleRevocation({ type: 'channel.follow', status: 'authorization_revoked', condition: { broadcaster_user_id: 'uid-revoke' } });
-
-    expect(clearStreamerToken).toHaveBeenCalledWith(100);
-    expect(logMock.warn).toHaveBeenCalledWith('Cleared token for revokedStreamer (authorization_revoked)');
-  });
-
-  it('clears the token when status is user_removed, using to_broadcaster_user_id', () => {
-    handleRevocation({ type: 'channel.raid', status: 'user_removed', condition: { to_broadcaster_user_id: 'uid-revoke' } });
-
-    expect(clearStreamerToken).toHaveBeenCalledWith(100);
-    expect(logMock.warn).toHaveBeenCalledWith('Cleared token for revokedStreamer (user_removed)');
-  });
-
-  it('does not clear the token for other revocation statuses', () => {
-    handleRevocation({ type: 'channel.follow', status: 'moderator_removed', condition: { broadcaster_user_id: 'uid-revoke' } });
-
-    expect(clearStreamerToken).not.toHaveBeenCalled();
-  });
-
-  it('logs an error if clearing the token itself fails', async () => {
-    vi.mocked(clearStreamerToken).mockRejectedValueOnce(new Error('db down'));
-
-    handleRevocation({ type: 'channel.follow', status: 'authorization_revoked', condition: { broadcaster_user_id: 'uid-revoke' } });
-    await new Promise((resolve) => setImmediate(resolve));
-
-    expect(logMock.error).toHaveBeenCalledWith('Clear token error:', expect.any(Error));
-  });
-});
-
-// ---------------------------------------------------------------------------
-// Subscription-group alert-type coverage (exhaustiveness)
-//
-// Guards against a new AlertEventType being added (to ALERT_EVENT_TYPES / the DB schema) without
-// also wiring it into a SUBSCRIPTION_GROUPS entry's `alertTypes` — a gap that would otherwise
-// compile fine and only surface at runtime as "the EventSub subscription is never created, so
-// the alert silently never fires."
-// ---------------------------------------------------------------------------
-describe('subscription-group alert-type coverage', () => {
-  it('covers every ALERT_EVENT_TYPES member in some subscription group', () => {
-    const covered = getAlertTypesCoveredBySubscriptionGroups();
-    for (const type of ALERT_EVENT_TYPES) {
-      expect(covered.has(type)).toBe(true);
-    }
-  });
-
-  it('does not cover any type outside ALERT_EVENT_TYPES (catches stale/typo\'d entries)', () => {
-    const covered = getAlertTypesCoveredBySubscriptionGroups();
-    for (const type of covered) {
-      expect(ALERT_EVENT_TYPES).toContain(type);
-    }
   });
 });
