+++ src/twitch/eventsub/twitchEventSubSubscriptions.ts
@@ -1,25 +1,17 @@
 import { createLogger } from '../../shared/logger';
-import { getAllEventSubStreamers, clearStreamerToken, getEnabledAlertEventTypesBatch, DEFAULT_EVENT_CONFIG } from '../../db';
+import { getAllEventSubStreamers, getEnabledAlertEventTypesBatch } from '../../db';
 import type { DbStreamerEventSub, EventSubConfig, AlertEventType } from '../../db';
 import { getUsers } from '../twitchApi';
 import { getActiveChannels } from '../twitchChannelMembership';
 import { normalizeTwitchChannelName } from '../twitchChannelName';
 import { createEventSubSubscription, listEventSubSubscriptions, deleteEventSubSubscription, getValidToken, TwitchAuthError } from './twitchApiEventSub';
-import {
-  handleFollow, handleSub, handleResub, handleGiftSub, handleRaid, handleRedemption,
-  handleStreamOnline, handleStreamOffline, handleChannelUpdate,
-  FollowEvent, SubEvent, ResubEvent, GiftSubEvent, RaidEvent, RedemptionEvent,
-} from './twitchEventSubHandler';
+import { SubSpec, SUBSCRIPTION_GROUPS, isGroupEnabled } from './twitchEventSubSubscriptionGroups';
+import { setStreamerInfo } from './twitchEventSubDispatch';
 
 const log = createLogger('EventSub');
 
-
-/** Describes a single EventSub subscription to create. */
-export interface SubSpec { type: string; version: string; condition: Record<string, string> }
-
-/** In-memory streamer info keyed by broadcaster user ID. */
-export interface StreamerInfo { login: string; streamerId: number; config: EventSubConfig | null }
-const streamerMap = new Map<string, StreamerInfo>();
+export { dispatchNotification, handleRevocation, removeStreamerFromMap } from './twitchEventSubDispatch';
+export { getAlertTypesCoveredBySubscriptionGroups } from './twitchEventSubSubscriptionGroups';
 
 // Tracks "login:type:token" triples that failed with 403 — skipped until bot restarts or token changes
 const authFailedSubs = new Set<string>();
@@ -36,21 +28,6 @@ export function clearAuthFailedSubs(login: string): void {
   const prefix = `${login}:`;
   for (const key of authFailedSubs) if (key.startsWith(prefix)) authFailedSubs.delete(key);
 }
-
-// Maps EventSub notification types to their handler functions.
-// Using Map instead of a plain object prevents prototype-chain lookup on user-controlled keys.
-type NotificationHandler = (login: string, event: unknown, config: EventSubConfig, streamerId: number) => Promise<void>;
-const notificationHandlers = new Map<string, NotificationHandler>([
-  ['channel.follow',                                   (l, e, c, sid) => handleFollow(l, e as FollowEvent, c, sid)],
-  ['channel.subscribe',                                (l, e, c, sid) => handleSub(l, e as SubEvent, c, sid)],
-  ['channel.subscription.message',                     (l, e, c, sid) => handleResub(l, e as ResubEvent, c, sid)],
-  ['channel.subscription.gift',                        (l, e, c, sid) => handleGiftSub(l, e as GiftSubEvent, c, sid)],
-  ['channel.raid',                                     (l, e, c, sid) => handleRaid(l, e as RaidEvent, c, sid)],
-  ['channel.channel_points_custom_reward_redemption.add',  (l, e, c, sid) => handleRedemption(l, e as RedemptionEvent, c, sid)],
-  ['stream.online',                                    (l) => handleStreamOnline(l)],
-  ['stream.offline',                                   (l) => handleStreamOffline(l)],
-  ['channel.update',                                    (l) => handleChannelUpdate(l)],
-]);
 
 /**
  * Deletes subscriptions that shouldn't be active any more: those whose type isn't in
@@ -103,117 +80,6 @@ async function resolveBroadcasterId(streamer: DbStreamerEventSub, config: EventS
     log.error(`Failed to resolve Twitch user ID for ${streamer.twitch_name}:`, err);
     return null;
   }
-}
-
-/** One gated group of EventSub subscriptions: created together whenever `isGroupEnabled` passes. */
-interface SubscriptionGroup {
-  /**
-   * Alert event types that share this group's subscription(s) — declaring these as data (rather
-   * than each group hand-writing its own `enabledAlerts.has('literal')` checks) lets
-   * `getAlertTypesCoveredBySubscriptionGroups` verify every {@link AlertEventType} is wired to a
-   * group; a test in `twitchEventSubSubscriptions.test.ts` fails if a new alert type is added
-   * without adding it here, instead of the gap silently compiling.
-   */
-  alertTypes: AlertEventType[];
-  /**
-   * Returns true if this group's subscriptions should be created based on the streamer's config
-   * alone — either a genuine chat-message toggle, or (for a group with no `alertTypes`)
-   * {@link hasCompletedEventSubSetup} as a proxy for "this streamer is set up enough to want this
-   * subscription at all". Independent of `alertTypes` — `isGroupEnabled` ORs the alerts-overlay
-   * gating in generically, since a streamer can want the subscription created purely to drive a
-   * browser-source alert, with chat messages off.
-   * @param config - The streamer's event response configuration, or null if unset.
-   */
-  configGate: (config: EventSubConfig | null) => boolean;
-  /**
-   * Builds this group's subscription specs.
-   * @param uid - The broadcaster's Twitch user ID.
-   * @returns The subscription specs to create for this group.
-   */
-  specs: (uid: string) => SubSpec[];
-}
-
-/**
- * Returns true if `group`'s subscriptions should be created: either its `configGate` passes, or
- * any of its `alertTypes` has an enabled alerts-overlay config.
- * @param group - The subscription group to evaluate.
- * @param config - The streamer's event response configuration, or null if unset.
- * @param enabledAlerts - Event types with an enabled alerts-overlay config row for this streamer.
- */
-function isGroupEnabled(
-  group: SubscriptionGroup,
-  config: EventSubConfig | null,
-  enabledAlerts: ReadonlySet<AlertEventType>,
-): boolean {
-  return group.configGate(config) || group.alertTypes.some((type) => enabledAlerts.has(type));
-}
-
-/**
- * `configGate` for a group with no `alertTypes` — unlike the alert-driven groups, there's no
- * alert-only path that needs the subscription without a real config row, so a row's mere
- * existence is used as a proxy for "this streamer completed the full EventSub/chat-message
- * setup". (`dispatchNotification` itself no longer early-exits without config — it falls back to
- * `DEFAULT_EVENT_CONFIG` for alert-only streamers — but that's moot for these groups since they
- * never subscribe for one.)
- * @param config - The streamer's event response configuration, or null if unset.
- */
-function hasCompletedEventSubSetup(config: EventSubConfig | null): boolean {
-  return Boolean(config);
-}
-
-// Every group also requires a broadcaster token (WebSocket transport only works with a user
-// token, not an app token — see createSubscriptionsForStreamer's upfront `!token` check).
-const SUBSCRIPTION_GROUPS: SubscriptionGroup[] = [
-  {
-    alertTypes: ['follow'],
-    configGate: (config) => Boolean(config?.follow_enabled),
-    specs: (uid) => [{ type: 'channel.follow', version: '2', condition: { broadcaster_user_id: uid, moderator_user_id: uid } }],
-  },
-  {
-    alertTypes: ['sub', 'resub', 'giftsub'],
-    configGate: (config) => Boolean(config?.sub_enabled),
-    specs: (uid) => [
-      { type: 'channel.subscribe', version: '1', condition: { broadcaster_user_id: uid } },
-      { type: 'channel.subscription.message', version: '1', condition: { broadcaster_user_id: uid } },
-      { type: 'channel.subscription.gift', version: '1', condition: { broadcaster_user_id: uid } },
-    ],
-  },
-  {
-    // Subscribe when the welcome message, the auto-shoutout toggle, or the raid alert is on —
-    // handleRaid gates its three behaviours independently, but the subscription itself must
-    // exist for any of them to fire.
-    alertTypes: ['raid'],
-    configGate: (config) => Boolean(config?.raid_enabled || config?.raid_shoutout_enabled),
-    specs: (uid) => [{ type: 'channel.raid', version: '1', condition: { to_broadcaster_user_id: uid } }],
-  },
-  {
-    alertTypes: [],
-    configGate: hasCompletedEventSubSetup,
-    specs: (uid) => [{ type: 'channel.channel_points_custom_reward_redemption.add', version: '1', condition: { broadcaster_user_id: uid } }],
-  },
-  {
-    // stream.online/offline and channel.update require no scope beyond a valid token. This
-    // drives an immediate live-check that supplements (not replaces) the 60s poller. Not tied to
-    // any alert type.
-    alertTypes: [],
-    configGate: hasCompletedEventSubSetup,
-    specs: (uid) => [
-      { type: 'stream.online', version: '1', condition: { broadcaster_user_id: uid } },
-      { type: 'stream.offline', version: '1', condition: { broadcaster_user_id: uid } },
-      { type: 'channel.update', version: '2', condition: { broadcaster_user_id: uid } },
-    ],
-  },
-];
-
-/**
- * Returns the set of alert event types covered by at least one subscription group. Exported for
- * the exhaustiveness test in `twitchEventSubSubscriptions.test.ts`, which asserts this covers
- * every member of `ALERT_EVENT_TYPES` — catching a new alert type being added without also being
- * wired into a subscription group's `alertTypes` (which would otherwise compile fine and only
- * surface as "the subscription is never created" at runtime).
- */
-export function getAlertTypesCoveredBySubscriptionGroups(): Set<AlertEventType> {
-  return new Set(SUBSCRIPTION_GROUPS.flatMap((group) => group.alertTypes));
 }
 
 /** Data bundle passed to a StreamerConnection for setting up EventSub subscriptions. */
@@ -289,21 +155,17 @@ export async function loadStreamersForEventSub(): Promise<StreamerEventSubData[]
   return result;
 }
 
-/** Creates all subscriptions for one streamer on their dedicated session, updates streamerMap,
- *  and cleans up stale subscriptions. Returns the count of desired subscriptions. */
+/** Creates all subscriptions for one streamer on their dedicated session, updates the
+ *  dispatch-side streamer map, and cleans up stale subscriptions. Returns the count of
+ *  desired subscriptions. */
 export async function subscribeForStreamer(
   sessionId: string, data: StreamerEventSubData,
 ): Promise<number> {
   const { uid, token, name, config, streamerId } = data;
-  streamerMap.set(uid, { login: name, streamerId, config });
+  setStreamerInfo(uid, { login: name, streamerId, config });
   const { desired, created } = await createSubscriptionsForStreamer(sessionId, data);
   await deleteStaleSubscriptions(uid, desired, created, token);
   return desired.size;
-}
-
-/** Removes a streamer from the in-memory map (called when their connection is stopped). */
-export function removeStreamerFromMap(uid: string): void {
-  streamerMap.delete(uid);
 }
 
 /** Creates a single EventSub subscription. Returns the created subscription's id, or null if
@@ -328,55 +190,5 @@ async function subscribe(sessionId: string, spec: SubSpec, token: string, login:
       log.error(`Failed to subscribe to ${spec.type} for ${login}:`, err);
     }
     return null;
-  }
-}
-
-
-/**
- * Routes an EventSub notification to the appropriate handler based on subscription type.
- * A streamer with no `streamer_event_config` row (`info.config` null) — e.g. one who has only
- * ever enabled a browser-source alert and never completed the chat-message OAuth setup — still
- * dispatches, using `DEFAULT_EVENT_CONFIG` (every chat-message flag disabled) in place of the
- * missing config, so their alert-only subscriptions aren't silently discarded here. Without
- * this fallback, `isGroupEnabled`'s alert-driven subscription creation and this dispatch gate
- * would disagree: the subscription would exist but every notification for it would be dropped.
- */
-export function dispatchNotification(type: string, event: Record<string, unknown>, condition: Record<string, string>): void {
-  const broadcasterId = condition.broadcaster_user_id ?? condition.to_broadcaster_user_id;
-  if (!broadcasterId) return;
-
-  const info = streamerMap.get(broadcasterId);
-  if (!info) return;
-  const config = info.config ?? DEFAULT_EVENT_CONFIG;
-
-  const handler = notificationHandlers.get(type);
-  if (!handler) {
-    log.warn(`Unsupported EventSub notification type: ${type}`);
-    return;
-  }
-  // TypeScript has already narrowed handler to NotificationHandler here, but the explicit
-  // typeof check is required for CodeQL's taint analysis to trust the call is safe.
-  if (typeof handler !== 'function') {
-    log.warn(`Invalid EventSub handler for type: ${type}`);
-    return;
-  }
-  handler(info.login, event, config, info.streamerId)
-    .catch((err) => log.error(`${type} handler error:`, err));
-}
-
-/** Handles a subscription revocation message, clearing the broadcaster's token if authorisation was revoked. */
-export function handleRevocation(sub: { type: string; status: string; condition: Record<string, string> }): void {
-  log.warn(`Subscription revoked: type=${sub.type} status=${sub.status}`);
-
-  if (sub.status === 'authorization_revoked' || sub.status === 'user_removed') {
-    const broadcasterId = sub.condition.broadcaster_user_id ?? sub.condition.to_broadcaster_user_id;
-    if (broadcasterId) {
-      const info = streamerMap.get(broadcasterId);
-      if (info) {
-        clearStreamerToken(info.streamerId)
-          .catch((err) => log.error('Clear token error:', err));
-        log.warn(`Cleared token for ${info.login} (${sub.status})`);
-      }
-    }
   }
 }
